Move Stage's useEffects triggered on annotation & selectedCategory to sagas
There were 2 useEffects which included `annotator` as a dependency (but did not list any of the other state dependencies it had causing a warning). This is better served as a side effect handled by saga. Likewise for the useEffect having selectedCategory as a sole dependency.

Create sagas for these state change events.