Draw mask border only after scaling
When producing an annotation mask, the mask is produced with respect to
the image's dimensions. The mask's (1px) border was then made opaque, while
the interior was made semi-transparent. The mask was then passed to
ImageJs along with a scaling factor, which scaled the image mask
according to the stage scale.

This resulted in the 1px border also scaling. Instead we want to first
scale the mask, then draw a 1px border.