Increase specificity of dependencies in useEffect
The useBoundingClientRect had a useEffect which included image as a
dependency.
This dispatched stage scaling whenever the image changes.
The image object has an annotation array.
So whenever the annotation array was pushed to, the stage would rescale.
Increase specificity of the dependency to only fire when the image shape
changes.